Improved and tested assertions and Logger

+++ common/debug.cc
@@ -3,7 +3,7 @@
 #if DEBUG_LEVEL != 0
 namespace boost
 {
-  void assertion_failed_msg(const char *expr, const char *msg, const char *function, const char *file, int line) noexcept {
+  void assertion_failed_msg(const char *expr, const char *msg, const char *function, const char *file, long line) noexcept {
     // Assertions are sent for logging, then application is aborted.
     // We assume assertion fail is a critical error.
     std::stringstream ss;
@@ -15,7 +15,7 @@ namespace boost
     Application::instance().Abort();
   }
 
-  void assertion_failed(const char *expr, const char *function, const char *file, int line) noexcept {
+  void assertion_failed(const char *expr, const char *function, const char *file, long line) noexcept {
     assertion_failed_msg(expr, NULL, function, file, line);
   }
 }
